requirements.txt
# PDF Parsing — stack leggero, stabile su Windows
# Usiamo pdfplumber per il parsing strutturato invece di unstructured[pdf]
# (evita la dipendenza da unstructured_inference/detectron2/tesseract)
pdfplumber>=0.11.0
pdfminer.six>=20221105

# LlamaIndex core + components
# 0.12.x è la serie stabile più recente (compatibile Python 3.13)
llama-index-core>=0.12.0
llama-index-readers-file>=0.4.0
llama-index-embeddings-huggingface>=0.4.0

# Vector DB
chromadb>=0.6.0

# Embeddings
sentence-transformers>=3.3.0
torch>=2.3.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
tqdm>=4.66.4